Chore: Health Connect 관련 화면 로직 변경
