+++ app/src/main/AndroidManifest.xml
@@ -50,6 +50,11 @@
             </intent-filter>
         </activity>
         <activity
+            android:name=".presentation.intro.HealthConnectActivity"
+            android:exported="true"
+            android:theme="@android:style/Theme.NoTitleBar">
+        </activity>
+        <activity
             android:name=".presentation.signup.SignupActivity"
             android:exported="true"
             android:theme="@android:style/Theme.NoTitleBar">
