+++ presentation/src/main/res/layout/activity_healthconnect.xml
